docs: add comprehensive SEO implementation guide with meta tags, sitemap and structured data
