+++ tests/Feature/Http/Controllers/CompanyControllerTest.php
@@ -5,10 +5,12 @@ namespace Tests\Feature\Http\Controllers;
 use App\Http\Controllers\CompanyController;
 use App\Models\Alternative;
 use App\Models\Company;
+use App\Notification\ReviewAlternative;
+use Illuminate\Support\Facades\Notification;
 
 use function Pest\Faker\fake;
+use function Pest\Laravel\from;
 use function Pest\Laravel\get;
-use function Pest\Laravel\post;
 
 mutates(CompanyController::class);
 
@@ -58,33 +60,54 @@ test('company is not shown if no approved', function (): void {
 });
 
 test('store an alternative inside company show', function (): void {
-    config()->set('honeypot.enabled', false);
+    Notification::fake();
 
+    // Arrange
     $company = Company::factory()->approved()->create();
+    $name = fake()->company();
+    $url = 'https://example.com';
 
-    $alternativeName = fake()->company();
-    $alternativeUrl = fake()->url();
+    // Act
+    $response = from(route('companies.show', $company->slug))
+        ->post(route('companies.alternatives.store', $company->slug), [
+            'name' => $name,
+            'url' => $url,
+        ]);
 
-    $response = post(route('companies.alternatives.store', $company->slug), [
-        'name' => $alternativeName,
-        'url' => $alternativeUrl,
-    ], [
-        'HTTP_REFERER' => route('companies.show', $company->slug),
-    ]);
+    // Assert
+    $alternatives = Alternative::query()
+        ->whereRelation('companies', 'id', $company->id)
+        ->where('name', $name)
+        ->where('url', $url)
+        ->get();
+
+    expect($alternatives)->toHaveCount(1);
+
+    $alternative = $alternatives->first();
 
     $this->assertDatabaseHas('alternatives', [
-        'name' => $alternativeName,
-        'url' => $alternativeUrl,
+        'name' => $name,
+        'url' => $url,
     ]);
 
     $this->assertDatabaseHas('alternative_company', [
         'company_id' => $company->id,
-        'alternative_id' => Alternative::query()->first()->id,
+        'alternative_id' => $alternative->id,
     ]);
 
     $response->assertRedirect(route('companies.show', $company->slug));
-
     $response->assertSessionHas('success', 'Thank you for suggesting an alternative');
+
+    /** @var User $adminUser * */
+    $adminUser = $this->adminUser;
+
+    Notification::assertSentTo(
+        $adminUser,
+        ReviewAlternative::class,
+        function ($notification) use ($alternative, $company): bool {
+            return $notification->alternative->is($alternative)
+                && $notification->company->is($company);
+        });
 });
 
 // test('store saves and redirects', function (): void {
